+++ Tests/Listener/RestListenerPathConverterTest.php
@@ -2,6 +2,8 @@
 
 namespace Paysera\Bundle\RestBundle\Tests;
 
+use Mockery;
+use Mockery\MockInterface;
 use Paysera\Bundle\RestBundle\ApiManager;
 use Paysera\Bundle\RestBundle\Exception\ApiException;
 use Paysera\Bundle\RestBundle\Listener\RestListener;
@@ -20,6 +22,7 @@ use Paysera\Component\Serializer\Factory\ContextAwareNormalizerFactory;
 use Paysera\Component\Serializer\Normalizer\ArrayNormalizer;
 use Paysera\Component\Serializer\Normalizer\ViolationNormalizer;
 use Paysera\Component\Serializer\Validation\PropertyPathConverterInterface;
+use PHPUnit_Framework_TestCase;
 use Psr\Log\NullLogger;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\ParameterBag;
@@ -29,26 +32,28 @@ use Symfony\Component\Validator\ConstraintViolation;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
-class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
+class RestListenerPathConverterTest extends PHPUnit_Framework_TestCase
 {
-    /** @var \Mockery\MockInterface|FilterControllerEvent */
+    /**
+     * @var MockInterface|FilterControllerEvent
+     */
     private $filterControllerEvent;
 
     public function setUp()
     {
-        $this->filterControllerEvent = \Mockery::mock(FilterControllerEvent::class);
+        $this->filterControllerEvent = Mockery::mock(FilterControllerEvent::class);
     }
 
     public function testOnKernelControllerWithRequestQueryMapperValidationThrowsExceptionWithCamelCasePathConverter()
     {
-        $exceptionThrowed = false;
+        $exceptionThrown = false;
         try {
             $this
                 ->createRestListener(new CamelCaseToSnakeCaseConverter())
                 ->onKernelController($this->filterControllerEvent)
             ;
         } catch (ApiException $apiException) {
-            $exceptionThrowed = true;
+            $exceptionThrown = true;
             $this->assertEquals(
                 [
                     'first_name' => ['firstName message'],
@@ -66,16 +71,17 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
             );
         }
 
-        $this->assertTrue($exceptionThrowed);
+        $this->assertTrue($exceptionThrown);
     }
 
     public function testOnKernelControllerWithRequestQueryMapperValidationThrowsExceptionWithNoOpConverter()
     {
-        $exceptionThrowed = false;
+        $exceptionThrown = false;
         try {
             $this->createRestListener(new NoOpConverter())->onKernelController($this->filterControllerEvent);
+            $this->setExpectedException(ApiException::class);
         } catch (ApiException $apiException) {
-            $exceptionThrowed = true;
+            $exceptionThrown = true;
             $this->assertEquals(
                 [
                     'firstName' => ['firstName message'],
@@ -93,10 +99,12 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
             );
         }
 
-        $this->assertTrue($exceptionThrowed);
+        $this->assertTrue($exceptionThrown);
     }
 
     /**
+     * @param PropertyPathConverterInterface $pathConverter
+     *
      * @return RestListener
      */
     private function createRestListener(PropertyPathConverterInterface $pathConverter)
@@ -114,7 +122,7 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
 
         $queryParameterBag->add($entity);
 
-        $request = \Mockery::mock(Request::class);
+        $request = Mockery::mock(Request::class);
         $request->shouldReceive('getContent')->andReturn('{}');
         $request->attributes = $parameterBag;
         $request->query = $queryParameterBag;
@@ -122,9 +130,9 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
         $this->filterControllerEvent->shouldReceive('getRequest')->andReturn($request);
 
         if (interface_exists('Symfony\Component\Validator\Validator\ValidatorInterface')) {
-            $validator = \Mockery::mock(ValidatorInterface::class);
+            $validator = Mockery::mock(ValidatorInterface::class);
         } else {
-            $validator = \Mockery::mock(LegacyValidatorInterface::class);
+            $validator = Mockery::mock(LegacyValidatorInterface::class);
         }
 
         $violationList = new ConstraintViolationList([
@@ -134,14 +142,14 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
 
         $validator->shouldReceive('validate')->andReturn($violationList);
 
-        $formatDetector = \Mockery::mock(FormatDetector::class);
+        $formatDetector = Mockery::mock(FormatDetector::class);
         $formatDetector->shouldReceive('getRequestFormat')->andReturn('json');
 
-        $requestMapper = \Mockery::mock(NameAwareDenormalizerInterface::class);
+        $requestMapper = Mockery::mock(NameAwareDenormalizerInterface::class);
         $requestMapper->shouldReceive('mapToEntity')->andReturn([]);
         $requestMapper->shouldReceive('getName')->andReturn('name');
 
-        $container = \Mockery::mock(ContainerInterface::class);
+        $container = Mockery::mock(ContainerInterface::class);
         $container->shouldReceive('get')->andReturn($requestMapper);
 
         $api = new RestApi($container, new NullLogger());
@@ -162,16 +170,17 @@ class RestListenerPathConverterTest extends \PHPUnit_Framework_TestCase
         $apiManager->addApiByKey($api, 'api');
         $apiManager->addDecoder(new Json(), 'json');
 
-        $parameterToEntityMapBuilder = \Mockery::mock(ParameterToEntityMapBuilder::class);
+        $parameterToEntityMapBuilder = Mockery::mock(ParameterToEntityMapBuilder::class);
         $parameterToEntityMapBuilder->shouldReceive('buildParameterToEntityMap')->andReturn([]);
 
         return new RestListener(
             $apiManager,
-            \Mockery::mock(ContextAwareNormalizerFactory::class),
+            Mockery::mock(ContextAwareNormalizerFactory::class),
             new NullLogger(),
             $parameterToEntityMapBuilder,
             new RequestLogger(new NullLogger()),
-            new ExceptionLogger()
+            new ExceptionLogger(),
+            []
         );
     }
 }